pyproject.toml
[tool.poetry]
name = "nano"
version = "3.8.3"
description = ""
authors = ["Simon Goričar <simon.peter.goricar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-rapidjson = "^0.9.4"
giphypop = "^0.3"
psutil = "^5.7.3"
beautifulsoup4 = "^4.9.3"
redis = "^3.5.3"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.12.0"
pillow = "^8.0.1"
"discord.py" = "^1.5.1"
tmdbie = {git = "https://github.com/DefaultSimon/TMDbie.git"}
osu-ds = {git = "https://github.com/DefaultSimon/osu-ds.git"}
steamapi = {git = "https://github.com/smiley/steamapi.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
